Extract types to src/types directory
