+++ src/types/arp-params.ts
@@ -1,0 +1,7 @@
+interface ArpParams {
+    count: number;
+    order?: string;
+    chords: string | any[];
+};
+
+export type { ArpParams };
