+++ src/types/note-object.ts
@@ -1,0 +1,7 @@
+interface NoteObject {
+    note: string[];
+    length: number;
+    level: number;
+}
+
+export type { NoteObject };
