+++ src/types/nvp.ts
@@ -1,0 +1,5 @@
+interface NVP<T> {
+    [key: string]: T;
+}
+
+export type { NVP };
