+++ src/types/player-observer-fn.ts
@@ -1,0 +1,7 @@
+/**
+ * Callback function triggered when a note is played.
+ * @param params - object with player data
+ */
+type PlayerObserverFn = (params: any) => void;
+
+export type { PlayerObserverFn };
