+++ src/types/progression-scale.ts
@@ -1,0 +1,3 @@
+type ProgressionScale = 'major' | 'minor' | 'M' | 'm';
+
+export type { ProgressionScale };
