+++ src/types/seq-fn.ts
@@ -1,0 +1,3 @@
+type SeqFn = (time: string, el: string) => void;
+
+export type { SeqFn };
