+++ src/types/sizzle-style.ts
@@ -1,0 +1,3 @@
+type SizzleStyle = 'sin' | 'cos' | 'rampUp' | 'rampDown';
+
+export type { SizzleStyle };
