+++ src/types/tpd.ts
@@ -1,0 +1,17 @@
+
+interface TPD {
+    /**
+     * Tonic
+     */
+    T: string[];
+    /**
+     * Predominant (or subdominant)
+     */
+    P: string[];
+    /**
+     * Dominant
+     */
+    D: string[];
+}
+
+export type { TPD };
